Move `CCCamera.cpp/h` `CCLight.cpp/h` into 2d
