+++ cocos/2d/CCCamera.h
