+++ cocos/2d/CCLight.h
