+++ cocos/2d/CMakeLists.txt
@@ -20,6 +20,7 @@ set(COCOS_2D_SRC
   2d/CCAnimationCache.cpp
   2d/CCAnimation.cpp
   2d/CCAtlasNode.cpp
+  2d/CCCamera.cpp
   2d/CCClippingNode.cpp
   2d/CCClippingRectangleNode.cpp
   2d/CCComponentContainer.cpp
@@ -43,6 +44,7 @@ set(COCOS_2D_SRC
   2d/CCLabelTextFormatter.cpp
   2d/CCLabelTTF.cpp
   2d/CCLayer.cpp
+  2d/CCLight.cpp
   2d/CCMenu.cpp
   2d/CCMenuItem.cpp
   2d/CCMotionStreak.cpp
