+++ cocos/2d/libcocos2d.vcxproj.filters
@@ -727,9 +727,6 @@
     <ClCompile Include="CCProtectedNode.cpp">
       <Filter>2d</Filter>
     </ClCompile>
-    <ClCompile Include="..\base\CCCamera.cpp">
-      <Filter>base</Filter>
-    </ClCompile>
     <ClCompile Include="..\renderer\CCPrimitive.cpp">
       <Filter>renderer</Filter>
     </ClCompile>
@@ -1198,9 +1195,6 @@
     <ClCompile Include="..\ui\UIWebView.cpp">
       <Filter>ui\UIWidgets</Filter>
     </ClCompile>
-    <ClCompile Include="..\base\CCLight.cpp">
-      <Filter>base</Filter>
-    </ClCompile>
     <ClCompile Include="..\audio\AudioEngine.cpp">
       <Filter>audioengine</Filter>
     </ClCompile>
@@ -1264,6 +1258,12 @@
     </ClCompile>
     <ClCompile Include="..\..\external\protobuf-lite\src\google\protobuf\stubs\stringprintf.cc">
       <Filter>cocostudio\json\protobuf-lite\stubs</Filter>
+    </ClCompile>
+    <ClCompile Include="CCCamera.cpp">
+      <Filter>2d</Filter>
+    </ClCompile>
+    <ClCompile Include="CCLight.cpp">
+      <Filter>2d</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -1841,9 +1841,6 @@
     <ClInclude Include="CCProtectedNode.h">
       <Filter>2d</Filter>
     </ClInclude>
-    <ClInclude Include="..\base\CCCamera.h">
-      <Filter>base</Filter>
-    </ClInclude>
     <ClInclude Include="..\renderer\CCPrimitive.h">
       <Filter>renderer</Filter>
     </ClInclude>
@@ -2414,9 +2411,6 @@
     <ClInclude Include="..\ui\UIWebView.h">
       <Filter>ui\UIWidgets</Filter>
     </ClInclude>
-    <ClInclude Include="..\base\CCLight.h">
-      <Filter>base</Filter>
-    </ClInclude>
     <ClInclude Include="..\audio\include\AudioEngine.h">
       <Filter>audioengine</Filter>
     </ClInclude>
@@ -2528,6 +2522,12 @@
     </ClInclude>
     <ClInclude Include="..\..\external\protobuf-lite\src\google\protobuf\stubs\type_traits.h">
       <Filter>cocostudio\json\protobuf-lite\stubs</Filter>
+    </ClInclude>
+    <ClInclude Include="CCCamera.h">
+      <Filter>2d</Filter>
+    </ClInclude>
+    <ClInclude Include="CCLight.h">
+      <Filter>2d</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
